refactoring back

+++ api/src/routes/controllers/deleteControllers.js
@@ -1,9 +1,15 @@
 const { Player, Event, Group, Product } = require("../../db");
+const rgExp = /^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$/;
+
 
 const asyncDeleteProduct = async (req, res) => {
   try {
-    await Product.destroy({ where: { id: req.params.id } });
-    return res.status(200).send("The product has been successfully removed");
+    if (rgExp.test(req.params.id)) {
+      await Product.destroy({ where: { id: req.params.id } });
+    res.status(200).json({message:"The product has been successfully removed"});
+  }else{
+      res.status(401).json({message:"id invalid"});
+    }
   } catch (error) {
     console.log(error);
   }
@@ -12,19 +18,21 @@ const asyncDeleteProduct = async (req, res) => {
 const deleteGroups = async (req, res) => {
   const { id } = req.params;
   try {
-    if (!id || id.length < 36) {
-      res
-        .json({ message: "id is require or id is to short, please try again" })
-        .status(400);
+    if (!id) {
+      res.status(400).json({ message: "id is require or id is to short, please try again" })
     } else {
-      const validateGroup = await Group.findByPk(id);
-      if (validateGroup !== null) {
-        await Group.destroy({
-          where: { id },
-        });
-        res.json({ message: "Group has been delete" }).status(200);
-      } else {
-        res.json({ message: "group not found" }).status(404);
+      if(rgExp.test(id)){
+        const validateGroup = await Group.findByPk(id);
+        if (validateGroup !== null) {
+          await Group.destroy({
+            where: { id },
+          });
+          res.status(200).json({ message: "Group has been delete" });
+        } else {
+          res.status(404).json({ message: "group not found" });
+        }
+      }else{
+        res.status(412).json({ message: "id invalid" });
       }
     }
   } catch (error) {
@@ -36,12 +44,16 @@ const deleteEvent = async (req, res) => {
   const { id } = req.params;
   try {
     if (!id) {
-      res.status(400).json({ error: "missing id" });
+      res.status(411).json({ error: "id is missing" });
     } else {
-      await Event.destroy({
-        where: { id },
-      });
-      res.json({ message: "Event deleted" });
+      if(rgExp.test(id)){
+        await Event.destroy({
+          where: { id },
+        });
+        res.json({ message: "Event deleted" });
+      }else{
+        res.status(400).json({ message: "Bad request" });
+      }
     }
   } catch (error) {
     res.status(400).json({ error_DB: error.message });
